gpt_pipe/gpt_requirements.txt
boto3==1.34.34
python-dotenv==1.0.1
openai==1.11.1
asyncio==3.4.3